# feat: Complete SELECT query optimization - achieve 16x overhead target for cached queries