… CLAUDE.md
… README.md
… TODO.md
… benchmarks/benchmark.py
… docs/SELECT_OPTIMIZATION_PLAN.md
… src/query/extended.rs
… src/query/fast_path.rs
… src/session/db_handler.rs
… tests/benchmark_cache_effectiveness.rs
… tests/query_cache_test.rs